fix: code style optimiz
