+++ web/public/index.html
@@ -53,7 +53,7 @@
     <script>
       less = {
         async: false,
-        env: 'development',
+        env: "%NODE_ENV%",
         javascriptEnabled: true
       };
     </script>
