+++ web/src/App.less
@@ -64,7 +64,7 @@
   }
 }
 
-.theme_btn {
+.themeBox {
   background-position: center !important;
   background-repeat: no-repeat !important;
   width: 40px;
@@ -88,8 +88,4 @@
   margin: 5px;
   flex: 1;
   align-items: stretch;
-}
-
-.head-wrap-card {
-  box-shadow: 0 1px 5px 0 rgb(51 51 51 / 14%);
 }
